+++ inst/shiny-examples/LakeMonitoR/ui.R
@@ -19,12 +19,14 @@
 tab_Help   <- source("external/tab_Help.R", local = TRUE)$value
 tab_Import <- source("external/tab_Import.R", local = TRUE)$value
 tab_Calc   <- source("external/tab_Calc.R", local = TRUE)$value
+tab_Plot   <- source("external/tab_Plot.R", local = TRUE)$value
 
 # Define UI
 shinyUI(
-  navbarPage("LakeMonitoR, v0.1.0.9008"
+  navbarPage("LakeMonitoR, v0.1.0.9018"
     , tab_Help()
     , tab_Import()
     , tab_Calc()
+    , tab_Plot()
   )##navbarPage ~ END
 )##shinyUI~END